+++ pkarr/src/extra/mod.rs
@@ -1,5 +1,3 @@
-use std::sync::Arc;
-
 #[cfg(feature = "endpoints")]
 pub mod endpoints;
 
@@ -14,18 +12,113 @@ pub mod lmdb_cache;
 
 #[cfg(all(not(target_arch = "wasm32"), feature = "reqwest-builder"))]
 impl From<crate::Client> for ::reqwest::ClientBuilder {
+    /// Create a [reqwest::ClientBuilder][::reqwest::ClientBuilder] from this Pkarr client,
+    /// using it as a [dns_resolver][::reqwest::ClientBuilder::dns_resolver],
+    /// and a [preconfigured_tls][::reqwest::ClientBuilder::use_preconfigured_tls] client
+    /// config that uses [rustls::crypto::ring::default_provider()] and follows the
+    /// [tls for pkarr domains](https://pkarr.org/tls) spec.
     fn from(client: crate::Client) -> Self {
         ::reqwest::ClientBuilder::new()
-            .dns_resolver(Arc::new(client.clone()))
+            .dns_resolver(std::sync::Arc::new(client.clone()))
             .use_preconfigured_tls(rustls::ClientConfig::from(client))
     }
 }
 
 #[cfg(all(not(target_arch = "wasm32"), feature = "reqwest-builder"))]
 impl From<crate::client::relay::Client> for ::reqwest::ClientBuilder {
+    /// Create a [reqwest::ClientBuilder][::reqwest::ClientBuilder] from this Pkarr client,
+    /// using it as a [dns_resolver][::reqwest::ClientBuilder::dns_resolver],
+    /// and a [preconfigured_tls][::reqwest::ClientBuilder::use_preconfigured_tls] client
+    /// config that uses [rustls::crypto::ring::default_provider()] and follows the
+    /// [tls for pkarr domains](https://pkarr.org/tls) spec.
     fn from(client: crate::client::relay::Client) -> Self {
         ::reqwest::ClientBuilder::new()
-            .dns_resolver(Arc::new(client.clone()))
+            .dns_resolver(std::sync::Arc::new(client.clone()))
             .use_preconfigured_tls(rustls::ClientConfig::from(client))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use mainline::Testnet;
+    use std::net::SocketAddr;
+    use std::net::TcpListener;
+    use std::sync::Arc;
+
+    use axum::{routing::get, Router};
+    use axum_server::tls_rustls::RustlsConfig;
+
+    use crate::{
+        dns::{rdata::SVCB, Packet},
+        Client, Keypair, SignedPacket,
+    };
+
+    async fn publish_server_pkarr(client: &Client, keypair: &Keypair, socket_addr: &SocketAddr) {
+        let mut packet = Packet::new_reply(1);
+
+        let mut svcb = SVCB::new(0, ".".try_into().unwrap());
+
+        svcb.set_port(socket_addr.port());
+
+        packet.answers.push(crate::dns::ResourceRecord::new(
+            "@".try_into().unwrap(),
+            crate::dns::CLASS::IN,
+            60 * 60,
+            crate::dns::rdata::RData::HTTPS(svcb.into()),
+        ));
+
+        packet.answers.push(crate::dns::ResourceRecord::new(
+            "@".try_into().unwrap(),
+            crate::dns::CLASS::IN,
+            60 * 60,
+            match socket_addr.ip() {
+                std::net::IpAddr::V4(ip) => crate::dns::rdata::RData::A(ip.into()),
+                std::net::IpAddr::V6(ip) => crate::dns::rdata::RData::AAAA(ip.into()),
+            },
+        ));
+
+        let signed_packet = SignedPacket::from_packet(&keypair, &packet).unwrap();
+
+        client.publish(&signed_packet).await.unwrap();
+    }
+
+    #[tokio::test]
+    async fn reqwest_pkarr_domain() {
+        let testnet = Testnet::new(3).unwrap();
+
+        let keypair = Keypair::random();
+
+        {
+            // Run a server on Pkarr
+            let app = Router::new().route("/", get(|| async { "Hello, world!" }));
+            let listener = TcpListener::bind("127.0.0.1:0").unwrap(); // Bind to any available port
+            let address = listener.local_addr().unwrap();
+
+            let client = Client::builder().testnet(&testnet).build().unwrap();
+            publish_server_pkarr(&client, &keypair, &address).await;
+
+            println!("Server running on https://{}", keypair.public_key());
+
+            let server = axum_server::from_tcp_rustls(
+                listener,
+                RustlsConfig::from_config(Arc::new((&keypair).into())),
+            );
+
+            tokio::spawn(server.serve(app.into_make_service()));
+        }
+
+        // Client setup
+        let pkarr_client = Client::builder().testnet(&testnet).build().unwrap();
+        let reqwest = reqwest::ClientBuilder::from(pkarr_client).build().unwrap();
+
+        // Make a request
+        let response = reqwest
+            .get(format!("https://{}", keypair.public_key()))
+            .send()
+            .await
+            .unwrap();
+
+        assert_eq!(response.status(), reqwest::StatusCode::OK);
+        assert_eq!(response.text().await.unwrap(), "Hello, world!");
     }
 }